Created 'Rango' application
Prior to this commit, we only had a shell of a project. Arguably,
it was the administration interface that could allow us to
interact with future application models BEFORE we add forms
to manipulate data.

In this chapter (04), the authors are stressing that we should
create an application for every core features of our site so that
each "application" can be re-used elsewhere if necessary.

We will try to keep that in mind going forward.

    --- Lionel Saliou, Ph.D