Enabling the 1st Rango view
This was punishing!

So much has changed in Django 2.0 that old syntax do not work.
Furthermore, the stuff leart on Pluralsight courses are difficult
to apply here too. For example, the "url" keyword does not apply
anymore. It has been replaced by "path" for simple URLs
and re_path when regular expression (with named place holders!)
are needed.

It was hard to debug. Perhaps because it was my first major error,
however it is telling of the difficulties going forward.

Since the code is so young, it might be worth it to change the
requirements and swap for an earlier version of Django.

That being said, the version the tutorial references is so old
that it uses "incorrect" methods. So, it might be a good exercise
to continue as such. Furthermore, with practice I should be able
to see through these problems.

    --- Lionel Saliou, Ph.D