Chapter 4 vows and Linter
Added a brief overview of the difficulties of following this
tutorial with respect to the advancement of DJANGO 2.0.

Also, added a Linter for Markdown and had to add a new configuration
file for this. I do not want this part of the repo so I modified
gitignore.

For reference, the .markdownlint.json file is currently as follows:
{
    "MD013": false,
    "MD029": { "style": "ordered" }
}

    --- Lionel Saliou, Ph.D