Use a RENDERER instead of HttpResponse object
This required adding a 'templates' directory to our application.
Note that, again, this is different in DJANGO 2.0 than it is
in the version outlined in the tutorial. So, here special attention
was given to the newer docs.

It must be noted that, by default, the DJANGO 2.0 project is setup
in such a way that the recommended way works almost by default.
Except for creating the folders and the template file, I had
nothing in terms of configuration to do. Thus, skipping a full
section of Chapter 05.

    --- Lionel Saliou, Ph.D