Using static image and namespacing
Again, I had to go back to the 2.0 documentations. As with templates,
it is recommended to create a sub directory of the name of the
application in the STATIC folder. This required a change in
the 'src' value of the IMG tag. Indeed, the 'static' keyword
allows for going through all static assets (without configuration
so far), however to find the correct asset, I needed to specify
from which application the images must come from. It was a simple
change...

    --- Lionel Saliou, Ph.D