Failed attempt to run in non-debug mode
I vaguely remember this topic being touched on in one of the
courses I have watched. However, here I am failing to fix the
problem. When running with DEBUG=FALSE, no amount of additional
configuration allows for the image to be shown. Indeed, in
debug mode, the 'static' folder is handled/served by the light
weight web-server. But not in 'standard' mode.

I have found snippets online of stuff I can add to the URLPATTERNS
array. However, everyone says this is not efficient. I will put
this aside for now. There is a point where this should be
tested/ran on a some online platform. Hopefully, it will work
just fine then.

    --- Lionel Saliou, Ph.D