Failed attempt to use the MEDIA concept
A lot of application allow users to upload artefacts and
especially images to use as avatars. During development,
such capabilities must be built and tested. This part of the
exercise was supposed to help ciment that concept. However,
I was met with failure. The file I have added to the media
folder cannot be shown ... so far. I must go back to the
code school video.

 --- Lionel Saliou, Ph.D