Fixed Media folder and success getting GIF file
I had to look at the Digging into Django code from code school
in order to make it work. As shown in this diff, I also had
to move the media folder to the root of the project.

On the one hand it makes sense. The application, such as Rango,
itself should not care where the 'media' are and they certainly
should not be tracked as part of the development.

   --- Lionel Saliou, Ph.D

Edit: forgot to delete the commented out line.