Completed Chapter 05 exercise
It was a bit frustrating to use the MEDIA folder in these
conditions. I believe the big error I did intially in the new
template is try to use to the %static% keyword. MEDIA may be
the location of static content but is not served through the
same means. Hence to get the Mr Geen Icon to show, I had to
use the same address I have been testing MEDIA with since the
last commit. In other words, I "pretend" that the resource
I need is hosted UNDER the application in its media folder...
even though on Disk this folder is elsewhere.

   --- Lionel Saliou, Ph.D