Moved MEDIA and STATIC conf to root URLs.py
It is not the responsibility of our application to specify and/or
manage the manner in which MEDIA and STATIC content is served;
debug mode or not.

In this commit, we move this code to the root configuration where
it belongs. Along with this, we correct the mistake of trying
to specify the "application" name as part of the path of files
in the media directory.

A quick test showed that static content is still being displayed
correctly and that items from media also appear provided the
URL is correct.

     --- Lionel Saliou, Ph.D