Display 'top categories' in order on Rango page
In this commit, we query the database for data related to our
Category model. Using the OO aspect of python, in one statement
we get all of the categories stored so far and place the top
5 in our context supplied to the template context.

In the template, we iterate throught the list (provided it is
valid) and display the results.

    --- Lionel Saliou, Ph.D