Added Slugfield to Category model
I understand the principle behind this. We want the URLs to be
human readable. However, I am not too sure when to it arise
during the design that this technique would be a solution.

I admit that the application I wish to develop next is quite
hazy... so no panic :-)

    --- Lionel Saliou, Ph.D