Fix routing URLs for Rango
In the process of testing the display of categories, I noticed
that it was not possible to display these new pages because
the 'matching' capabilities were to broad.

As I would still like for the About/Index link to work, I concluded
that the root application URLs.py had to cope with both the root
request and Rango specific requests.

Likewise, I had forgotten that in DJANGO 2.0 the matching is much
simplified. There is no need to use the caret (^) for example.

Now the routing is working ok.

    --- Lionel Saliou, Ph.D