Category form created
Perhaps, it is tiredness however this felt a bit more difficult
than the previous chapter.

I added to the add_category template both the form suggested in
the exercise as well as the 'as_p' method. The latter's output
is not so good.

There are two exercises to do in this chapter. The first one
is likely to require a new 'template' to say to the user
that the category already exists. I would prefer some pop-up
in Javascript to write out a simple message but I do not know
how to do that (yet)!

     --- Lionel Saliou, Ph.D