Use base template and content blocks!
I assumed I would need to create yet another page to show that
the category already existed. Therefore, it grew tiresome
to copy and paste the same HTML base code all the time.

Using the Digging into Django Treasure gram app as an inspiration,
I created a base template and derived all existing pages from it.

I already had a simple CSS file from before (PhotoUploader) and
I used that to ensure that the footer displayed better. Along with,
this I can see that multiple content blocks can be used. Nice!

Now, there will be no need to create a new page to handle the
condition where the category already exists. We will simply
show a DIV with a message to this effect.

    --- Lionel Saliou, Ph.D