Completed Chapter 08 exercise
This took much longer than I ever wanted.
1 - during testing I discovered that slugs are case insensitive
and therefore a Category CSharp and csharp resolve to the same
slug and this causes problems when visiting the category's page.
2 - to resolve this I had to implement a 'clean' function
in the form for the category model
3 - use the same approach in the add_page as in add_category
in terms of how errors are returned to the user
4 - discovered that the URLField does not accept urls starting
with HTTPS! I had to modify the clean method to fix this. However
more testing reveale that www.example.com would not be accepted.
This seems to be by default because putting debug statements in
clean were not triggered.

     --- Lionel Saliou, Ph.D