Add registration with Rango using multiple models
In this commit, we expand from the standard DJANGO user model
by having additional fields. This requires use to implement
two new models and these lead to two new forms that are handled
within the same request. The key here is that one of the model's
handling is delayed, eg commit is set to False.

Testing showed that the new pictures are indeed uploaded to the
media folder in a dedicated sub-folder (thanks Pillow!). However,
since this data is created during testing, the content therein
is not tracked in GIT.

   --- Lionel Saliou, Ph.D