Outlined idea for error handling in base template
Indeed, right now the errors are printed back in a dedicated
DIV block but for EACH form page. Yes, it will be rare for
a non-form page to print error, but a more universal approach
in the base template will lower code duplication in HTML pages
and allow for consistency in the views.

    --- Lionel Saliou, Ph.D