Adjust links presence based on page & user status
This showed again the power of using the template language.
In this commit, we leverage the `base.html` file to show
the Login and Register links according to user status.
Alon with this, we hidde the redundant links, eg no need to show
the login link on the login page!

    --- Lionel Saliou, Ph.D