Chapter 09 exercises completed
Ensured that adding category or page requires user to be logged in.
Refactored form pages so that error handling is 'displayed' in
an universal fashion from the base.html page. This technique
supports multiple 'forms' errors such as it can be experienced
with user registration.

The only "snag" is the slug error when the category already
exists. The display starts with __ALL__ which is not pretty.

   --- Lionel Saliou, Ph.D