Use namespace in template URL
As noted in the Django documentation, it could very well be that
'indexes' or 'details' pages exist for every "app" that compose
a site. Therefore, there must be a way of mapping this correctly.
Here we use the "rango" namespace.

The server complained that it was not a unique namespace because
we imported it twice in the "project directory". I simplified
this URLs.py however, I had to expand Rango's because otherwise
some of the redirection (such as from the login page) would fail.

     --- Lionel Saliou, Ph.D