Chapter 11 exercise: visits on About page
This demonstrates that once the data is set in the session object,
on the index page for exmaple, then the data can be re-used else
where.

Although, it must be noted that the data to be used must be "copied"
in the context for the request. In other words, a vast amount of
data may be stored in the session object however it does not mean
that all is accessible to all pages. The view code is responsible
for this.

   --- Lionel Saliou, Ph.D