Added: Editor Config file but ignore vscode cfg
This is a best practice nowadays to ensure consistency across
developers and editors.

I created a vscode workspace for Rango and, for now, it is not
tracked. This workspace should make use of the .editorconfig.
At time of writing, I did not see any "warnings" such what
I have seen for the Python code with PyLint and yet I know for
a fact that some of the code I have written so far does not
conform to editorconfig specifications. These were taken
directly from the DJANGO github page.

    --- Lionel Saliou, Ph.D