Completed chapter 12
Quite a bit of work done in this chapter. We removed all of the
code we produced for login/loggout and regitration. We are
using a dedicated application instead.

On the plus side, this application is a dependancy that is simply
captured in requiremetns.txt for PIP.

On the downside, it would appear that depending where this
application is loaded in the settings.py file, the site breaks.
For example, if loaded as recommended before the AUTH package
from django, the login page fails.

It would also appear that this new application assumes where
the templates are located. As advised, I am using namespacing.
Perhaps, my settings are incorrect in terms of how DJANGO should
search for the templates.

Still, the only visible problem here is that the password change
form does not load ANY of the code we produced for this chapter.
This is strange given that other forms, such as login or registration
correctly load "our style".

    --- Lionel Saliou, Ph.D