FIXED password change page
When defining the TEMPLATE_DIR variable, it is possible to
force the display of our page provided that:
a) In TEMPLATES, we set the DIRS key to [TEMPLATES, ]
b) we create a templates/regitration folder in the
tango_with_django_project directory

In my view, this is bad. Indeed, the APP_DIRS is set to True, so
the framework should search for templates in our application
directories too! Right now, there is only 1 application, Rango.
However, it has templates for both itself and registration.

It does not make sense that Login and Registration for example
are found correctly.

EDIT: issue fixed

This was resolved by ensuring that RANGO is the first application
loaded. The template engine probably looks-up stuff in order.
Because the layout of the page was obviously inspired by the
style of the admin app, this suggested that there already is
a corresponding file. With Rango loaded first, the engine
finds the files we created. This is perhaps why the django-auth-redux
project suggests loading 'registration' before 'auth'. However
for us, it broke the application. Also, I checked and we are following
the recommended approach with template definition and storage. So, there
might still be bugs within the django-auth-redux application.

    --- Lionel Saliou, Ph.D