Chapter 13 completed
We installed the django-bootstrap4 application. Using its special
tags, we ensured that we could use the various CSS classes to
style our application. In this chapter, we added a pretty
nifty navigation bar at the top of the application and it is
fast too as it is probably entierely done in CSS.

Needless to say, I am pretty weak in terms of CSS. However, using
the bootstraps docs, it should be okay to style the site into
something a bit better.

The installation process was painless. The only difficulty was
to do go to the django-bootstrap4 github and grab the base
template. It is not enough to simply call the tag.

In conclusion, it is a powerful feature to be able to install
an application like this and quickly leverage the CSS code
a mature framework, such as bootstrap 4, has.

   --- Lionel Saliou, Ph.D

EDIT: ensure that bootstrap4 is captured in the requirements.txt