Chapter 14 Template Tags completed
This is quite an amazing chapter. It suggests that we can
develop a site in chuncks with each "panel" or areas being its
own logic.

I am quite impressed by Bootstraps also. The results from the
code suggested in the tutorial were poor for readability but
a quite google search yeild something much more of my tests.
Along with this, it was combined with inline Django Template Logic
that allowed for choosing the correct classes and make the overall
look better. Really neat!

Note: this chapter does not appear to be included in the "latest"
version of this tutorial.

    --- Lionel Saliou, Ph.D