Added ability to track page views
Added a view, mapped the URL, used the request content instead
of creating a slug as a parameter to the view. Leveraged the
existing page model to look-up the page url using its internal
id.

NOTE: there was a problem with the redirection in case the ID
was incorrect. If the redirection object is given the argument
of "rango/" it loops forever. This might be a combinaison
of things such as the fact that 'rango' keyword is used and
mapped in both the admin part and the rango application side.

   --- Lionel Saliou, Ph.D