Users fill-in their profile after registration
In this commit, we redirect users who have just registered
to a new form that allows them to fill-in the details for
their user profile, such as Website and avatar.

This proved challenging. First of all, we had to override the
default behaviour for the registration success view in the admin
site so that the redirection was the new form and not the index.

Second, we had difficulties overriding properly as the function
definition changed so that the request was no longer a required
parameter. The stack trace was not understood quickly. It required
looking into the actual source code for django-auth-redux to figure
out!

Third, the new UserProfile class and model were not added to the
admin.py and therefore, no data could be seen when testing the
new workflow!

Fourth, the profile_registration requires work. Rigth now we
use the shorthand from Bootstrap4 to get the form to render whereas
in other pages, we create, place and style each field/buttons
individually.

NOTE: we need to fix the way forms delegated to django-auth-redux
provide user feedback. Right now, the page simply reloads. The
data is lost and the user does not know what went wrong.

    --- Lionel Saliou, Ph.D