Acknowledging CSS rubbishness
Wrote down officially how much time was wasted today trying to fix
element positioning and so on with almost no CSS knowledge.

Also, looked into providing the login ability directly in the top
nav-bar. Again, the position of the form is not to my liking
and lots of time wasted trying to fix something I do not fully
understand. An "acceptable" design is captured in top_nav_bar.html
and it is planned that his will be a templatetags soon so that
the nav bar content can be changed according to where on the site
the user is and whether he/she is authenticated. This done at the
moment in the base.html file in both the header and footer section.
It is hoped that the logic can be factored out.

    --- Lionel Saliou, Ph.D