Cope with user without avatars
In this commit, we attempted to use glythicons. However, this is
no longer supported in Bootstrap 4 due to copyright and licensing
issues. Instead, we use Font Awesome. It uses the same 'approach'
as bootstrap3 did so it was not too difficult to adapt to our
needs.

By accident, this commit also includes the required code to
ensure that only user him/herself can edit their own profile.

    --- Lionel Saliou, Ph.D