Support User List and Profile edit
Whilst creating the view to list all of the users with valid
profiles was fast and easy, the update of existing user profiles
proved more difficult.

It is the binding of existing records to the data coming into the
request that proved the hardest to debug. This is something that
the tutorial does not cover at all, so a lot of reading to achieve
this result.

    --- Lionel Saliou, Ph.D