Allow user to reset image without deleting default
Through this commit, we learn how to add a new checkbox into a
form that is not part of the underlying Model the form is based on.
Despite this, it is possible to leverage the name/value this
additional control brings to drive some special logic.

In this instance, we decided to allow the users to reset their
avatar if they have not supplied their own and are not happy
with the one provided by default.

The key was to ensure that they would not delete the default.jpg
file in the process. This is working.

The one snag is that typically when the user changes avatar
the old file is removed from disk. If they decide to reset instead
this does not happen. So orphan files are left on disk.

This concludes successfully chapter 16!

    --- Lionel Saliou, Ph.D