Partially adopt django-bootstrap4 error handling
This is not functional yet. In other words if I type the wrong
password in a form, there is no data being displayed in this
area of the form. However, it is quite neat with the cross
on the right to dismiss each and every message.

Perhaps, I will have to re-work the error handling system in order
to take advantage of it. Most of the forms are not within Rango
though so it might be harder to wire them correctly.

    --- Lionel Saliou, Ph.D