Added ability to LIKE a category
WARNING: always develop with script blockers DISABLED!!!!
This last commit should have been easy but the requests were
failing for no apparent reasons. All the logs suggested that
the server was not even seeing the request coming through.

Using Google Chrome almost work immediately! Firefox and Security
gah! the rage!

For future reference, the Javascript code is kept verbose and
documented. The python views are very simple and easy to maintain.
There was no need to return JSON structures for now.

SNAG: the user can re-like a category after refreshing or coming
to visit the category page at a later point in time...

    --- Lionel Saliou, Ph.D