Implemented category suggestion
Now that AJAX works, this was quite quick to implement. There
were bugs of course but nothing that could not be overcome.
As always, the code recieved additional documentation for the
future.

As we near the end of Chapter 19 I see that there is one
exercise we cannot complete. In other words, there is no
search function in Rango at the moment and therefore we cannot
implelement the feature where we add a page searched for
easily using AJAX.

    --- Lionel Saliou, Ph.D