BUG FIX: allow 'category' in category name
As a point of curiosity I created a category named
'A New Category' to ensure that the slug was coping with multiple
spaces and so on. It was not possible to add a page to this
category! This was due to the manner in which the URL match
was written. It was matching on the last instance of the keyword
and thus assumed that 'add_page' was the category to display.

Worse it was matching the wrong path and thus displaying the wrong
page.

Sorted now :-)

    --- Lionel Saliou, Ph.D