Added tested functionality for first/last visits
Very interesting last exercise! Indeed, we had to deal with the
date time data type. This required to use a DJANGO specific class
to handle timezones data correctly.

Along with this, we respected the idea of a skinny view and fast
models. Indeed, the behaviour in terms of how the new dates, when
the user visits a tracked page, are dealt with is coded into the
model save() overload. No need for complex logic when dealing
with the request!

This is almost an end-to-end change as it is was required to
migrate the existing model to cope with the new fieldds. We also
changed the HTML template to smartly show or hide the visit
information.

It took longer than I wanted but it was a worthwhile change.
Especially since I developed the tests before the model's logic.

    --- Lionel Saliou, Ph.D